feat: allow selecting text in epub reader
